+++ src/main/java/com/project/project/controller/UserController.java
@@ -21,6 +21,12 @@ public class UserController {
 
     @GetMapping(path = "/admin/user/addnew")
     public String addNewUser(Model model){
+        if (!userService.isUserLogin()) {
+            return "redirect:/login";
+        }
+        if (!userService.isUserRoleIsAdmin()){
+            return "redirect:/page_not_found";
+        }
         List<Role> roles = roleService.findAll();
         List<User> users = userService.findAll();
         model.addAttribute("roles", roles);
@@ -30,18 +36,36 @@ public class UserController {
 
     @PostMapping(path = "/admin/user/save")
     public String saveNewUser(User user){
+        if (!userService.isUserLogin()) {
+            return "redirect:/login";
+        }
+        if (!userService.isUserRoleIsAdmin()){
+            return "redirect:/page_not_found";
+        }
         userService.save(user);
         return "redirect:/admin/user";
     }
 
     @RequestMapping(path = "/admin/user/delete/{id}", method = {RequestMethod.GET, RequestMethod.DELETE})
     public String deleteUserById(@PathVariable Long id){
+        if (!userService.isUserLogin()) {
+            return "redirect:/login";
+        }
+        if (!userService.isUserRoleIsAdmin()){
+            return "redirect:/page_not_found";
+        }
         userService.deleteById(id);
         return "redirect:/admin/user";
     }
 
     @GetMapping(path = "/admin/user/update/{id}")
     public String findUserById(@PathVariable Long id, Model model){
+        if (!userService.isUserLogin()) {
+            return "redirect:/login";
+        }
+        if (!userService.isUserRoleIsAdmin()){
+            return "redirect:/page_not_found";
+        }
         User user = userService.findById(id).orElse(null);
         List<Role> roles = roleService.findAll();
         model.addAttribute("user", user);
